requirements.txt
python-docx==1.1.0
lxml==4.9.3